WideChar2Utf8 __int64

+++ phpwb.h
@@ -73,7 +73,7 @@ zval *process_array(zval *zitems);
 
 TCHAR *Utf82WideChar(const char *str, int len);
 void Utf82WideCharCopy(const char *str, int str_len, TCHAR *wcs, int wcs_len);
-char *WideChar2Utf8(LPCTSTR wcs, int *len);
+char *WideChar2Utf8(LPCTSTR wcs, __int64 *len);
 void WideCharCopy(LPCTSTR wcs, char *s, int len);
 void dumptcs(TCHAR *str);
 char *ConvertUTF16ToUTF8(LPCWSTR pszTextUTF16, int *plen);
